requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pydantic>=2.0.0
py-algorand-sdk>=2.0.0
python-dotenv>=1.0.0
